Push more actions down to the only components that use them
